refactor(frontend): split ModelPanel into hook + 3 sub-components (B-3)
Break the 484-line ModelPanel God component into an orchestrator, a data hook, and three focused presentation components. Behavior is preserved; all existing ModelPanel / ConfigForm tests stay green without mock updates.

Structure after the split:
- hooks/useModelPanelData.ts (~241 lines): 5 useQuery fan-out (schema / config / backends / uiSchema / columns), debounced validation on change, import / undo / redo / preset save+load handlers, and the derived fitEnabled / tuneEnabled / disabledReason state. 11 new unit tests cover the derived state priority order, validation debounce, import error surfacing, and the no-op paths (running, equal-to-cached).
- components/workspace/ModelPanelHeader.tsx (~100 lines): sticky tabs + primary Fit/Tune button + backend label + disabledReason hint.
- components/workspace/ConfigEditorBody.tsx (~115 lines): scrollable body hosting ConfigForm or TuneTab plus running / validation-error overlays. Preserves the WCAG 2.1.1 scrollable-region-focusable tabIndex (biome override extended to this file).
- components/workspace/ModelPanelActions.tsx (~166 lines): footer action bar (Import YAML / Export / Undo / Redo / Save Preset / Load Preset / Raw Config). Owns its own file input ref — no external coupling.
- components/workspace/ModelPanel.tsx (484 -> 106 lines): pure orchestrator — holds tab state + SavePresetDialog open state, wires the hook output into the three presentational children.

useConfigPresets.ts now exports the ConfigPreset interface so ModelPanelActions can type its presets prop without reaching into internal module state.

Stats: 1541 vitest (+11 new) pass, biome clean, pnpm build clean.

Known pre-existing issues (out of scope for this refactor, candidates for follow-up): handleLoadPreset emits a success toast before awaiting handleConfigChange; validateConfig catch silently leaves stale errors state. Both behaviors are preserved as-is from the pre-refactor ModelPanel.
